+++ src/aioway/hop/__init__.py
@@ -1,4 +1,6 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
+from .funcs import *
 from .hop import *
 from .nn import *
+from .tensors import *
